feat: restructure tests into package
