+++ core/tests/__init__.py
@@ -1,0 +1,4 @@
+from .test_admin_views import *
+from .test_forms import *
+from .test_parsing import *
+from .test_general import *
