+++ core/tests/test_general.py
@@ -6,7 +6,7 @@ from django.db import IntegrityError
 
 
 from django.apps import apps
-from .models import (
+from ..models import (
     BVProject,
     BVProjectFile,
     Recording,
@@ -29,7 +29,7 @@ from .models import (
     AntwortErkennungsRegel,
     Anlage4Config,
 )
-from .docx_utils import (
+from ..docx_utils import (
     extract_text,
     get_docx_page_count,
     get_pdf_page_count,
@@ -38,14 +38,14 @@ from .docx_utils import (
     _normalize_header_text,
 )
 
-from . import text_parser
+from .. import text_parser
 
 from core.text_parser import parse_anlage2_text, PHRASE_TYPE_CHOICES
 
-from .anlage4_parser import parse_anlage4
+from ..anlage4_parser import parse_anlage4
 
-from .parser_manager import parser_manager
-from .parsers import AbstractParser
+from ..parser_manager import parser_manager
+from ..parsers import AbstractParser
 
 from pathlib import Path
 from tempfile import NamedTemporaryFile
@@ -54,15 +54,15 @@ from PIL import Image
 import fitz
 
 from django.core.files.uploadedfile import SimpleUploadedFile
-from .forms import (
+from ..forms import (
     BVProjectForm,
     BVProjectUploadForm,
     BVProjectFileJSONForm,
     Anlage2ConfigForm,
 )
-from .workflow import set_project_status
-from .models import ProjectStatus
-from .llm_tasks import (
+from ..workflow import set_project_status
+from ..models import ProjectStatus
+from ..llm_tasks import (
     classify_system,
     check_anlage1,
     check_anlage2,
@@ -83,8 +83,8 @@ from .llm_tasks import (
     parse_anlage1_questions,
     _parse_anlage2,
 )
-from .views import _verification_to_initial
-from .reporting import generate_gap_analysis, generate_management_summary
+from ..views import _verification_to_initial
+from ..reporting import generate_gap_analysis, generate_management_summary
 from unittest.mock import patch, ANY
 from django.core.management import call_command
 from django.test import override_settings
@@ -131,7 +131,7 @@ def seed_test_data() -> None:
     create_initial_data = import_module(
         "core.migrations.0002_seed_initial_data"
     ).create_initial_data
-    from .llm_tasks import (
+    from ..llm_tasks import (
         ANLAGE1_QUESTIONS,
         _ANLAGE1_INTRO,
         _ANLAGE1_IT,
@@ -176,775 +176,7 @@ class NoesisTestCase(TestCase):
         )
 
 
-class AdminProjectsTests(NoesisTestCase):
-    def setUp(self):
-        admin_group = Group.objects.create(name="admin")
-        self.user = User.objects.create_user("admin", password="pass")
-        self.user.groups.add(admin_group)
-        self.client.login(username="admin", password="pass")
-
-        self.p1 = create_project(["A"], beschreibung="x")
-        self.p2 = create_project(["B"], beschreibung="y")
-
-    def test_delete_selected_projects(self):
-        url = reverse("admin_projects")
-        resp = self.client.post(url, {"delete_selected": "1", "selected_projects": [self.p1.id]})
-        self.assertRedirects(resp, url)
-        self.assertFalse(BVProject.objects.filter(id=self.p1.id).exists())
-        self.assertTrue(BVProject.objects.filter(id=self.p2.id).exists())
-
-    def test_delete_single_project(self):
-        url = reverse("admin_projects")
-        resp = self.client.post(url, {"delete_single": str(self.p2.id)})
-        self.assertRedirects(resp, url)
-        self.assertFalse(BVProject.objects.filter(id=self.p2.id).exists())
-
-    def test_delete_single_requires_post(self):
-        url = reverse("admin_project_delete", args=[self.p1.id])
-        resp = self.client.get(url)
-        self.assertEqual(resp.status_code, 405)
-        self.assertTrue(BVProject.objects.filter(id=self.p1.id).exists())
-
-
-class AdminProjectCleanupTests(NoesisTestCase):
-    def setUp(self):
-        admin_group = Group.objects.create(name="admin")
-        self.user = User.objects.create_user("admin2", password="pass")
-        self.user.groups.add(admin_group)
-        self.client.login(username="admin2", password="pass")
-
-        self.projekt = create_project(["A"], beschreibung="x")
-        self.file = BVProjectFile.objects.create(
-            projekt=self.projekt,
-            anlage_nr=1,
-            upload=SimpleUploadedFile("a.txt", b"data"),
-            text_content="Text",
-        )
-
-    def test_delete_file(self):
-        path = Path(self.file.upload.path)
-        url = reverse("admin_project_cleanup", args=[self.projekt.pk])
-        resp = self.client.post(url, {"action": "delete_file", "file_id": self.file.id})
-        self.assertRedirects(resp, url)
-        self.assertFalse(BVProjectFile.objects.filter(id=self.file.id).exists())
-        self.assertFalse(path.exists())
-
-    def test_delete_gutachten(self):
-        gpath = generate_gutachten(self.projekt.pk, text="foo")
-        url = reverse("admin_project_cleanup", args=[self.projekt.pk])
-        resp = self.client.post(url, {"action": "delete_gutachten"})
-        self.assertRedirects(resp, url)
-        self.projekt.refresh_from_db()
-        self.assertEqual(self.projekt.gutachten_file.name, "")
-        self.assertFalse(gpath.exists())
-
-    def test_delete_classification(self):
-        self.projekt.classification_json = {"a": 1}
-        self.projekt.save()
-        url = reverse("admin_project_cleanup", args=[self.projekt.pk])
-        resp = self.client.post(url, {"action": "delete_classification"})
-        self.assertRedirects(resp, url)
-        self.projekt.refresh_from_db()
         self.assertIsNone(self.projekt.classification_json)
-
-
-class DocxExtractTests(NoesisTestCase):
-    def test_extract_text(self):
-        doc = Document()
-        doc.add_paragraph("Das ist ein Test")
-        tmp = NamedTemporaryFile(delete=False, suffix=".docx")
-        doc.save(tmp.name)
-        tmp.close()
-        try:
-            text = extract_text(Path(tmp.name))
-        finally:
-            Path(tmp.name).unlink(missing_ok=True)
-        self.assertIn("Das ist ein Test", text)
-
-    def test_get_docx_page_count_single(self):
-        doc = Document()
-        doc.add_paragraph("Seite 1")
-        tmp = NamedTemporaryFile(delete=False, suffix=".docx")
-        doc.save(tmp.name)
-        tmp.close()
-        try:
-            count = get_docx_page_count(Path(tmp.name))
-        finally:
-            Path(tmp.name).unlink(missing_ok=True)
-        self.assertEqual(count, 1)
-
-    def test_get_docx_page_count_two_pages(self):
-        doc = Document()
-        doc.add_paragraph("Seite 1")
-        doc.add_page_break()
-        doc.add_paragraph("Seite 2")
-        tmp = NamedTemporaryFile(delete=False, suffix=".docx")
-        doc.save(tmp.name)
-        tmp.close()
-        try:
-            count = get_docx_page_count(Path(tmp.name))
-        finally:
-            Path(tmp.name).unlink(missing_ok=True)
-        self.assertEqual(count, 2)
-
-    def test_get_pdf_page_count_single(self):
-        pdf = fitz.open()
-        pdf.new_page()
-        tmp = NamedTemporaryFile(delete=False, suffix=".pdf")
-        pdf.save(tmp.name)
-        tmp.close()
-        try:
-            count = get_pdf_page_count(Path(tmp.name))
-        finally:
-            Path(tmp.name).unlink(missing_ok=True)
-        self.assertEqual(count, 1)
-
-    def test_get_pdf_page_count_two_pages(self):
-        pdf = fitz.open()
-        pdf.new_page()
-        pdf.new_page()
-        tmp = NamedTemporaryFile(delete=False, suffix=".pdf")
-        pdf.save(tmp.name)
-        tmp.close()
-        try:
-            count = get_pdf_page_count(Path(tmp.name))
-        finally:
-            Path(tmp.name).unlink(missing_ok=True)
-        self.assertEqual(count, 2)
-
-    def test_normalize_header_text_variants(self):
-        cases = {
-            "Technisch vorhanden?": "technisch vorhanden",
-            "Technisch vorhanden:" : "technisch vorhanden",
-            "Technisch   vorhanden": "technisch vorhanden",
-            "Technisch\tvorhanden": "technisch vorhanden",
-            " Verf\u00fcgbar?\t": "verf\u00fcgbar",
-        }
-        for raw, expected in cases.items():
-            self.assertEqual(_normalize_header_text(raw), expected)
-
-    def test_parse_anlage2_table(self):
-        doc = Document()
-        table = doc.add_table(rows=2, cols=5)
-        table.cell(0, 0).text = "Funktion"
-        table.cell(0, 1).text = "Technisch vorhanden"
-        table.cell(0, 2).text = "Einsatz bei Telefónica"
-        table.cell(0, 3).text = "Zur LV-Kontrolle"
-        table.cell(0, 4).text = "KI-Beteiligung"
-
-        table.cell(1, 0).text = "Login"
-        table.cell(1, 1).text = "Ja"
-        table.cell(1, 2).text = "Nein"
-        table.cell(1, 3).text = "Nein"
-        table.cell(1, 4).text = "Ja"
-
-        tmp = NamedTemporaryFile(delete=False, suffix=".docx")
-        doc.save(tmp.name)
-        tmp.close()
-        try:
-            with patch("core.docx_utils.logging.getLogger") as mock_get_logger:
-                mock_logger = mock_get_logger.return_value
-                data = parse_anlage2_table(Path(tmp.name))
-                expected_raw = [
-                    "Funktion",
-                    "Technisch vorhanden",
-                    "Einsatz bei Telefónica",
-                    "Zur LV-Kontrolle",
-                    "KI-Beteiligung",
-                ]
-                expected_norm = [
-                    "funktion",
-                    "technisch vorhanden",
-                    "einsatz bei telefónica",
-                    "zur lv-kontrolle",
-                    "ki-beteiligung",
-                ]
-                mock_logger.debug.assert_any_call(
-                    f"Tabelle 0: Roh-Header = {expected_raw}, Normiert = {expected_norm}"
-                )
-        finally:
-            Path(tmp.name).unlink(missing_ok=True)
-
-        self.assertEqual(
-            data,
-            [
-                {
-                    "funktion": "Login",
-                    "technisch_verfuegbar": {"value": True, "note": None},
-                    "einsatz_telefonica": {"value": False, "note": None},
-                    "zur_lv_kontrolle": {"value": False, "note": None},
-                    "ki_beteiligung": {"value": True, "note": None},
-                }
-            ],
-        )
-
-    def test_parse_anlage2_table_multiple_headers(self):
-        cfg = Anlage2Config.get_instance()
-        Anlage2ColumnHeading.objects.create(
-            config=cfg, field_name="technisch_vorhanden", text="Verfügbar?"
-        )
-        Anlage2ColumnHeading.objects.create(
-            config=cfg,
-            field_name="technisch_vorhanden",
-            text="Steht technisch zur Verfügung?",
-        )
-        Anlage2ColumnHeading.objects.create(
-            config=cfg, field_name="einsatz_bei_telefonica", text="Telefonica Einsatz"
-        )
-        Anlage2ColumnHeading.objects.create(
-            config=cfg, field_name="zur_lv_kontrolle", text="LV Kontrolle"
-        )
-        Anlage2ColumnHeading.objects.create(
-            config=cfg, field_name="ki_beteiligung", text="KI?"
-        )
-        doc = Document()
-        table = doc.add_table(rows=2, cols=5)
-        table.cell(0, 0).text = "Funktion"
-        table.cell(0, 1).text = "Steht technisch zur Verfügung?"
-        table.cell(0, 2).text = "Telefonica Einsatz"
-        table.cell(0, 3).text = "LV Kontrolle"
-        table.cell(0, 4).text = "KI?"
-        table.cell(1, 0).text = "Login"
-        table.cell(1, 1).text = "Ja"
-        table.cell(1, 2).text = "Nein"
-        table.cell(1, 3).text = "Nein"
-        table.cell(1, 4).text = "Ja"
-        tmp = NamedTemporaryFile(delete=False, suffix=".docx")
-        doc.save(tmp.name)
-        tmp.close()
-        try:
-            data = parse_anlage2_table(Path(tmp.name))
-        finally:
-            Path(tmp.name).unlink(missing_ok=True)
-
-        self.assertTrue(data[0]["technisch_verfuegbar"]["value"])
-
-    def test_parse_anlage2_table_alias_headers(self):
-        cfg = Anlage2Config.get_instance()
-        Anlage2ColumnHeading.objects.create(
-            config=cfg,
-            field_name="technisch_vorhanden",
-            text="Steht technisch zur Verfügung?",
-        )
-        Anlage2ColumnHeading.objects.create(
-            config=cfg,
-            field_name="einsatz_bei_telefonica",
-            text="einsatzweise bei telefónica: soll die funktion verwendet werden?",
-        )
-        Anlage2ColumnHeading.objects.create(
-            config=cfg,
-            field_name="zur_lv_kontrolle",
-            text="einsatzweise bei telefónica: soll zur überwachung von leistung oder verhalten verwendet werden?",
-        )
-
-        doc = Document()
-        table = doc.add_table(rows=2, cols=5)
-        table.cell(0, 0).text = "Funktion"
-        table.cell(0, 1).text = "Steht technisch zur Verfügung?"
-        table.cell(0, 2).text = "einsatzweise bei Telefónica: soll die Funktion verwendet werden?"
-        table.cell(0, 3).text = "einsatzweise bei Telefónica: soll zur Überwachung von Leistung oder Verhalten verwendet werden?"
-        table.cell(0, 4).text = "KI-Beteiligung"
-
-        table.cell(1, 0).text = "Login"
-        table.cell(1, 1).text = "Ja"
-        table.cell(1, 2).text = "Nein"
-        table.cell(1, 3).text = "Nein"
-        table.cell(1, 4).text = "Ja"
-
-        tmp = NamedTemporaryFile(delete=False, suffix=".docx")
-        doc.save(tmp.name)
-        tmp.close()
-        try:
-            data = parse_anlage2_table(Path(tmp.name))
-        finally:
-            Path(tmp.name).unlink(missing_ok=True)
-
-        self.assertEqual(
-            data,
-            [
-                {
-                    "funktion": "Login",
-                    "technisch_verfuegbar": {"value": True, "note": None},
-                    "einsatz_telefonica": {"value": False, "note": None},
-                    "zur_lv_kontrolle": {"value": False, "note": None},
-                    "ki_beteiligung": {"value": True, "note": None},
-                }
-            ],
-        )
-
-
-    def test_parse_anlage2_table_extra_text(self):
-        cfg = Anlage2Config.get_instance()
-        Anlage2ColumnHeading.objects.create(
-            config=cfg,
-            field_name="technisch_vorhanden",
-            text="Steht technisch zur Verf\u00fcgung?",
-        )
-        Anlage2ColumnHeading.objects.create(
-            config=cfg,
-            field_name="einsatz_bei_telefonica",
-            text="einsatzweise bei telef\u00f3nica: soll die funktion verwendet werden?",
-        )
-        Anlage2ColumnHeading.objects.create(
-            config=cfg,
-            field_name="zur_lv_kontrolle",
-            text="einsatzweise bei telef\u00f3nica: soll zur \u00fcberwachung von leistung oder verhalten verwendet werden?",
-        )
-
-        doc = Document()
-        table = doc.add_table(rows=2, cols=5)
-        table.cell(0, 0).text = "Funktion"
-        table.cell(0, 1).text = "Steht technisch zur Verf\u00fcgung?\n\nJa/nein"
-        table.cell(0, 2).text = (
-            "einsatzweise bei Telef\u00f3nica: soll die Funktion verwendet werden?\n\nJa/nein"
-        )
-        table.cell(0, 3).text = (
-            "einsatzweise bei Telef\u00f3nica: soll zur \u00dcberwachung von Leistung oder Verhalten verwendet werden?\n\nJa / nein"
-        )
-        table.cell(0, 4).text = "KI-Beteiligung"
-
-        table.cell(1, 0).text = "Login"
-        table.cell(1, 1).text = "Ja"
-        table.cell(1, 2).text = "Nein"
-        table.cell(1, 3).text = "Nein"
-        table.cell(1, 4).text = "Ja"
-
-        tmp = NamedTemporaryFile(delete=False, suffix=".docx")
-        doc.save(tmp.name)
-        tmp.close()
-        try:
-            data = parse_anlage2_table(Path(tmp.name))
-        finally:
-            Path(tmp.name).unlink(missing_ok=True)
-
-        self.assertEqual(
-            data,
-            [
-                {
-                    "funktion": "Login",
-                    "technisch_verfuegbar": {"value": True, "note": None},
-                    "einsatz_telefonica": {"value": False, "note": None},
-                    "zur_lv_kontrolle": {"value": False, "note": None},
-                    "ki_beteiligung": {"value": True, "note": None},
-                }
-            ],
-        )
-
-    def test_parse_anlage2_table_notes(self):
-        doc = Document()
-        table = doc.add_table(rows=2, cols=4)
-        table.cell(0, 0).text = "Funktion"
-        table.cell(0, 1).text = "Technisch vorhanden"
-        table.cell(0, 2).text = "Einsatz bei Telefónica"
-        table.cell(0, 3).text = "Zur LV-Kontrolle"
-
-        table.cell(1, 0).text = "Login"
-        table.cell(1, 1).text = "Ja (nur intern)"
-        table.cell(1, 2).text = "Nein, später"
-        table.cell(1, 3).text = "Nein (k.A.)"
-
-        tmp = NamedTemporaryFile(delete=False, suffix=".docx")
-        doc.save(tmp.name)
-        tmp.close()
-        try:
-            data = parse_anlage2_table(Path(tmp.name))
-        finally:
-            Path(tmp.name).unlink(missing_ok=True)
-
-        self.assertEqual(
-            data,
-            [
-                {
-                    "funktion": "Login",
-                    "technisch_verfuegbar": {"value": True, "note": "nur intern"},
-                    "einsatz_telefonica": {"value": False, "note": "später"},
-                    "zur_lv_kontrolle": {"value": False, "note": "k.A."},
-                }
-            ],
-        )
-
-    def test_parse_anlage2_table_alias_conflict(self):
-        cfg = Anlage2Config.get_instance()
-        conflict = "Konflikt"
-        Anlage2ColumnHeading.objects.create(
-            config=cfg, field_name="technisch_vorhanden", text=conflict
-        )
-        Anlage2ColumnHeading.objects.create(
-            config=cfg, field_name="einsatz_bei_telefonica", text=conflict
-        )
-
-        doc = Document()
-        table = doc.add_table(rows=2, cols=4)
-        table.cell(0, 0).text = "Funktion"
-        table.cell(0, 1).text = conflict
-        table.cell(0, 2).text = "Einsatz bei Telefónica"
-        table.cell(0, 3).text = "Zur LV-Kontrolle"
-
-        table.cell(1, 0).text = "Login"
-        table.cell(1, 1).text = "Ja"
-        table.cell(1, 2).text = "Nein"
-        table.cell(1, 3).text = "Nein"
-
-        tmp = NamedTemporaryFile(delete=False, suffix=".docx")
-        doc.save(tmp.name)
-        tmp.close()
-        try:
-            with self.assertRaises(ValueError):
-                parse_anlage2_table(Path(tmp.name))
-        finally:
-            Path(tmp.name).unlink(missing_ok=True)
-
-    def test_parse_anlage2_table_invalid_docx(self):
-        tmp = NamedTemporaryFile(delete=False, suffix=".docx")
-        tmp.write(b"invalid")
-        tmp.close()
-        try:
-            data = parse_anlage2_table(Path(tmp.name))
-        finally:
-            Path(tmp.name).unlink(missing_ok=True)
-
-        self.assertEqual(data, [])
-
-
-    def test_parse_anlage2_text(self):
-        func = Anlage2Function.objects.create(
-            name="Login",
-            detection_phrases={"name_aliases": ["login"]},
-        )
-        Anlage2SubQuestion.objects.create(
-            funktion=func,
-            frage_text="Warum?",
-            detection_phrases={"name_aliases": ["warum"]},
-        )
-        cfg = Anlage2Config.get_instance()
-        cfg.text_technisch_verfuegbar_true = ["tv ja"]
-        cfg.text_technisch_verfuegbar_false = ["tv nein"]
-        cfg.text_ki_beteiligung_false = ["ki nein"]
-        cfg.save()
-        text = "Login tv ja ki nein\nWarum? tv nein"
-        data = parse_anlage2_text(text)
-        self.assertEqual(
-            data,
-            [
-                {
-                    "funktion": "Login",
-                    "technisch_verfuegbar": {"value": True, "note": None},
-                    "ki_beteiligung": {"value": False, "note": None},
-                },
-                {
-                    "funktion": "Login: Warum?",
-                    "technisch_verfuegbar": {"value": False, "note": None},
-                },
-            ],
-        )
-
-    def test_parse_anlage2_text_default_aliases(self):
-        func = Anlage2Function.objects.create(name="Login")
-        Anlage2SubQuestion.objects.create(
-            funktion=func,
-            frage_text="Warum?",
-        )
-        cfg = Anlage2Config.get_instance()
-        cfg.text_technisch_verfuegbar_true = ["tv ja"]
-        cfg.text_technisch_verfuegbar_false = ["tv nein"]
-        cfg.text_ki_beteiligung_false = ["ki nein"]
-        cfg.save()
-        text = "Login tv ja ki nein\nWarum? tv nein"
-        data = parse_anlage2_text(text)
-        self.assertEqual(
-            data,
-            [
-                {
-                    "funktion": "Login",
-                    "technisch_verfuegbar": {"value": True, "note": None},
-                    "ki_beteiligung": {"value": False, "note": None},
-                },
-                {
-                    "funktion": "Login: Warum?",
-                    "technisch_verfuegbar": {"value": False, "note": None},
-                },
-            ],
-        )
-
-    def test_parse_anlage2_text_name_always_alias(self):
-        func = Anlage2Function.objects.create(
-            name="Anmelden",
-            detection_phrases={"name_aliases": ["login"]},
-        )
-        Anlage2SubQuestion.objects.create(
-            funktion=func,
-            frage_text="Grund?",
-            detection_phrases={"name_aliases": ["reason"]},
-        )
-        cfg = Anlage2Config.get_instance()
-        cfg.text_technisch_verfuegbar_true = ["tv ja"]
-        cfg.text_technisch_verfuegbar_false = ["tv nein"]
-        cfg.save()
-        text = "Anmelden tv ja\nGrund? tv nein"
-        data = parse_anlage2_text(text)
-        self.assertEqual(
-            data,
-            [
-                {
-                    "funktion": "Anmelden",
-                    "technisch_verfuegbar": {"value": True, "note": None},
-                },
-                {
-                    "funktion": "Anmelden: Grund?",
-                    "technisch_verfuegbar": {"value": False, "note": None},
-                },
-            ],
-        )
-
-    def test_parse_anlage2_text_normalizes_variants(self):
-        func = Anlage2Function.objects.create(name="User Login")
-        cfg = Anlage2Config.get_instance()
-        cfg.text_technisch_verfuegbar_true = ["tv ja"]
-        cfg.save()
-        text = "User-Login   tv ja"
-        data = parse_anlage2_text(text)
-        self.assertEqual(
-            data,
-            [
-                {
-                    "funktion": "User Login",
-                    "technisch_verfuegbar": {"value": True, "note": None},
-                }
-            ],
-        )
-
-    def test_parse_anlage2_text_punctuation_variants(self):
-        func = Anlage2Function.objects.create(
-            name="Analyse-/Reportingfunktionen",
-            detection_phrases={"name_aliases": ["Analyse-/Reportingfunktionen"]},
-        )
-        cfg = Anlage2Config.get_instance()
-        cfg.text_technisch_verfuegbar_true = ["tv ja"]
-        cfg.save()
-        text = "Analyse- / Reportingfunktionen tv ja"
-        data = parse_anlage2_text(text)
-        self.assertEqual(
-            data,
-            [
-                {
-                    "funktion": "Analyse-/Reportingfunktionen",
-                    "technisch_verfuegbar": {"value": True, "note": None},
-                }
-            ],
-        )
-
-    def test_parse_anlage2_text_prefers_specific_subquestion(self):
-        func = Anlage2Function.objects.create(name="Analyse-/Reportingfunktionen")
-        Anlage2SubQuestion.objects.create(
-            funktion=func,
-            frage_text="Bitte wähle zutreffendes aus",
-        )
-        cfg = Anlage2Config.get_instance()
-        cfg.text_technisch_verfuegbar_true = ["ja"]
-        cfg.save()
-        text = "Analyse- / Reportingfunktionen - Bitte wähle zutreffendes aus: ja"
-        data = parse_anlage2_text(text)
-        self.assertEqual(
-            data,
-            [
-                {
-                    "funktion": "Analyse-/Reportingfunktionen - Bitte wähle zutreffendes aus",
-                    "technisch_verfuegbar": {"value": True, "note": None},
-                }
-            ],
-        )
-
-    def test_parse_anlage2_text_merges_duplicate_functions(self):
-        func = Anlage2Function.objects.create(name="Login")
-        cfg = Anlage2Config.get_instance()
-        cfg.text_technisch_verfuegbar_true = ["tv ja"]
-        cfg.text_ki_beteiligung_false = ["ki nein"]
-        cfg.save()
-        text = "Login tv ja\nLogin ki nein"
-        data = parse_anlage2_text(text)
-        self.assertEqual(
-            data,
-            [
-                {
-                    "funktion": "Login",
-                    "technisch_verfuegbar": {"value": True, "note": None},
-                    "ki_beteiligung": {"value": False, "note": None},
-                }
-            ],
-        )
-
-    def test_parse_anlage2_text_updates_values_without_function(self):
-        func = Anlage2Function.objects.create(name="Analyse")
-        cfg = Anlage2Config.get_instance()
-        cfg.text_technisch_verfuegbar_true = ["verfuegbar"]
-        cfg.text_zur_lv_kontrolle_false = ["kein lv"]
-        cfg.save()
-        text = "Analyse verfuegbar\nkein lv"
-        data = parse_anlage2_text(text)
-        self.assertEqual(
-            data,
-            [
-                {
-                    "funktion": "Analyse",
-                    "technisch_verfuegbar": {"value": True, "note": None},
-                    "zur_lv_kontrolle": {"value": False, "note": None},
-                }
-            ],
-        )
-
-    def test_parse_anlage2_text_fuzzy_match(self):
-        func = Anlage2Function.objects.create(name="Login")
-        cfg = Anlage2Config.get_instance()
-        cfg.text_technisch_verfuegbar_true = ["ja"]
-        cfg.save()
-        data = parse_anlage2_text("Logn: ja")
-        self.assertEqual(
-            data,
-            [
-                {
-                    "funktion": "Login",
-                    "technisch_verfuegbar": {"value": True, "note": None},
-                }
-            ],
-        )
-
-    def test_parse_anlage2_text_multiple_rules_priority(self):
-        func = Anlage2Function.objects.create(name="Login")
-        AntwortErkennungsRegel.objects.create(
-            regel_name="a",
-            erkennungs_phrase="foo",
-            ziel_feld="technisch_verfuegbar",
-            wert=True,
-            prioritaet=2,
-        )
-        AntwortErkennungsRegel.objects.create(
-            regel_name="b",
-            erkennungs_phrase="bar",
-            ziel_feld="einsatz_telefonica",
-            wert=False,
-            prioritaet=1,
-        )
-        data = parse_anlage2_text("Login: foo bar rest")
-        self.assertEqual(
-            data,
-            [
-                {
-                    "funktion": "Login",
-                    "technisch_verfuegbar": {"value": True, "note": None},
-                    "einsatz_telefonica": {"value": False, "note": "rest"},
-                }
-            ],
-        )
-
-    def test_parse_anlage2_text_unknown_question(self):
-        Anlage2Function.objects.create(name="Login")
-        cfg = Anlage2Config.get_instance()
-        cfg.text_technisch_verfuegbar_true = ["ja"]
-        cfg.save()
-        data = parse_anlage2_text("Unbekannt: ja\nLogin: ja")
-        self.assertEqual(len(data), 1)
-        self.assertEqual(
-            data[0],
-            {
-                "funktion": "Login",
-                "technisch_verfuegbar": {"value": True, "note": None},
-            },
-        )
-
-class FormatBParserTests(NoesisTestCase):
-    def test_parse_format_b_basic(self):
-        text = "Login; tv: ja; tel: nein; lv: nein; ki: ja"
-        data = text_parser.parse_format_b(text)
-        self.assertEqual(
-            data,
-            [
-                {
-                    "funktion": "Login",
-                    "technisch_verfuegbar": {"value": True, "note": None},
-                    "einsatz_telefonica": {"value": False, "note": None},
-                    "zur_lv_kontrolle": {"value": False, "note": None},
-                    "ki_beteiligung": {"value": True, "note": None},
-                }
-            ],
-        )
-
-    def test_parse_format_b_numbering(self):
-        text = "1. Logout - tv=nein - ki=ja"
-        data = text_parser.parse_format_b(text)
-        self.assertEqual(
-            data,
-            [
-                {
-                    "funktion": "Logout",
-                    "technisch_verfuegbar": {"value": False, "note": None},
-                    "ki_beteiligung": {"value": True, "note": None},
-                }
-            ],
-        )
-
-    def test_parse_format_b_multiple_lines(self):
-        text = "Login; tv: ja\nLogout; tv: nein"
-        data = text_parser.parse_format_b(text)
-        self.assertEqual(len(data), 2)
-        self.assertTrue(data[0]["technisch_verfuegbar"]["value"])
-        self.assertFalse(data[1]["technisch_verfuegbar"]["value"])
-
-    def test_parse_format_b_with_rules(self):
-        FormatBParserRule.objects.create(key="t", target_field="technisch_verfuegbar")
-        text = "Login; t: ja"
-        data = text_parser.parse_format_b(text)
-        self.assertTrue(data[0]["technisch_verfuegbar"]["value"])
-
-    def test_extract_images(self):
-        img = Image.new("RGB", (1, 1), color="blue")
-        img_tmp = NamedTemporaryFile(delete=False, suffix=".png")
-        img.save(img_tmp.name)
-        img_tmp.close()
-        doc = Document()
-        doc.add_picture(img_tmp.name)
-        tmp = NamedTemporaryFile(delete=False, suffix=".docx")
-        doc.save(tmp.name)
-        tmp.close()
-        try:
-            data = extract_images(Path(tmp.name))
-        finally:
-            Path(tmp.name).unlink(missing_ok=True)
-            Path(img_tmp.name).unlink(missing_ok=True)
-        self.assertEqual(len(data), 1)
-        self.assertTrue(data[0].startswith(b"\x89PNG"))
-
-
-class BVProjectFormTests(NoesisTestCase):
-    def test_project_form_docx_validation(self):
-        data = QueryDict(mutable=True)
-        data.update(
-            {
-                "title": "",
-            }
-        )
-        data.setlist("software", ["A"])
-        form = BVProjectForm(data)
-        self.assertTrue(form.is_valid())
-        self.assertNotIn("docx_file", form.fields)
-
-    def test_upload_form_docx_validation(self):
-        valid = BVProjectUploadForm(
-            {}, {"docx_file": SimpleUploadedFile("t.docx", b"d")}
-        )
-        self.assertTrue(valid.is_valid())
-        invalid = BVProjectUploadForm(
-            {}, {"docx_file": SimpleUploadedFile("t.txt", b"d")}
-        )
-        self.assertFalse(invalid.is_valid())
-
-
-class Anlage2ConfigFormTests(NoesisTestCase):
-    def test_parser_order_field(self):
-        cfg = Anlage2Config.get_instance()
-        form = Anlage2ConfigForm({"parser_order": ["table"]}, instance=cfg)
-        self.assertTrue(form.is_valid())
-        inst = form.save()
-        self.assertEqual(inst.parser_order, ["table"])
 
 
 class BVProjectFileTests(NoesisTestCase):
@@ -2001,36 +1233,6 @@ class PromptTests(NoesisTestCase):
         self.assertEqual(p.text, expected)
 
 
-class AdminPromptsViewTests(NoesisTestCase):
-    def setUp(self):
-        admin_group = Group.objects.create(name="admin")
-        self.user = User.objects.create_user("padmin", password="pass")
-        self.user.groups.add(admin_group)
-        self.client.login(username="padmin", password="pass")
-        self.prompt = Prompt.objects.create(name="p1", text="orig")
-
-    def test_update_prompt(self):
-        url = reverse("admin_prompts")
-        resp = self.client.post(
-            url,
-            {
-                "pk": self.prompt.id,
-                "text": "neu",
-                "action": "save",
-                "use_system_role": "on",
-            },
-        )
-        self.assertRedirects(resp, url)
-        self.prompt.refresh_from_db()
-        self.assertEqual(self.prompt.text, "neu")
-
-    def test_delete_prompt(self):
-        url = reverse("admin_prompts")
-        resp = self.client.post(url, {"pk": self.prompt.id, "action": "delete"})
-        self.assertRedirects(resp, url)
-        self.assertFalse(Prompt.objects.filter(id=self.prompt.id).exists())
-
-
 class PromptImportTests(NoesisTestCase):
     def setUp(self):
         admin_group = Group.objects.create(name="admin")
@@ -2069,50 +1271,6 @@ class ReportingTests(NoesisTestCase):
         finally:
             path.unlink(missing_ok=True)
 
-    def test_management_summary_includes_comment(self):
-        projekt = BVProject.objects.create(software_typen="A", beschreibung="x")
-        BVProjectFile.objects.create(
-            projekt=projekt,
-            anlage_nr=1,
-            upload=SimpleUploadedFile("a.txt", b"data"),
-            text_content="Testtext",
-            manual_comment="Hinweis",
-        )
-        path = generate_management_summary(projekt)
-        try:
-            doc = Document(path)
-            text = "\n".join(p.text for p in doc.paragraphs)
-            self.assertIn("Hinweis", text)
-        finally:
-            path.unlink(missing_ok=True)
-
-    def test_manual_analysis_overrides(self):
-        projekt = BVProject.objects.create(software_typen="A", beschreibung="x")
-        BVProjectFile.objects.create(
-            projekt=projekt,
-            anlage_nr=1,
-            upload=SimpleUploadedFile("a.txt", b"data"),
-            text_content="Testtext",
-            analysis_json={"foo": {"value": "orig", "editable": True}},
-            manual_analysis_json={"foo": "manual"},
-        )
-        path1 = generate_gap_analysis(projekt)
-        try:
-            doc = Document(path1)
-            text = "\n".join(p.text for p in doc.paragraphs)
-            self.assertIn('"foo": "manual"', text)
-            self.assertNotIn('"foo": "orig"', text)
-        finally:
-            path1.unlink(missing_ok=True)
-
-        path2 = generate_management_summary(projekt)
-        try:
-            doc = Document(path2)
-            text = "\n".join(p.text for p in doc.paragraphs)
-            self.assertIn('"foo": "manual"', text)
-            self.assertNotIn('"foo": "orig"', text)
-        finally:
-            path2.unlink(missing_ok=True)
 
 
 class ProjektFileCheckViewTests(NoesisTestCase):
@@ -2189,125 +1347,6 @@ class ProjektFileCheckViewTests(NoesisTestCase):
             for i in nums
         }
         self.assertEqual(file_obj.analysis_json, expected)
-
-
-class ProjektFileJSONEditTests(NoesisTestCase):
-    def setUp(self):
-        self.user = User.objects.create_user("user3", password="pass")
-        self.client.login(username="user3", password="pass")
-        self.projekt = BVProject.objects.create(software_typen="A", beschreibung="x")
-        self.file = BVProjectFile.objects.create(
-            projekt=self.projekt,
-            anlage_nr=4,
-            upload=SimpleUploadedFile("a.txt", b"data"),
-            text_content="Text",
-            analysis_json={"old": {"value": True, "editable": True}},
-        )
-        self.anlage1 = BVProjectFile.objects.create(
-            projekt=self.projekt,
-            anlage_nr=1,
-            upload=SimpleUploadedFile("b.txt", b"data"),
-            text_content="Text",
-            analysis_json={
-                "questions": {
-                    "1": {
-                        "answer": "foo",
-                        "status": None,
-                        "hinweis": "",
-                        "vorschlag": "",
-                    }
-                }
-            },
-        )
-
-    def test_edit_json_updates_and_reports(self):
-        url = reverse("projekt_file_edit_json", args=[self.file.pk])
-        resp = self.client.post(
-            url,
-            {
-                "analysis_json": '{"new": 1}',
-                "manual_analysis_json": '{"manual": 2}',
-            },
-        )
-        self.assertEqual(resp.status_code, 302)
-        self.file.refresh_from_db()
-        self.assertEqual(self.file.analysis_json["new"], 1)
-        self.assertEqual(self.file.manual_analysis_json["manual"], 2)
-        path = generate_gap_analysis(self.projekt)
-        try:
-            doc = Document(path)
-            text = "\n".join(p.text for p in doc.paragraphs)
-            self.assertIn('"manual": 2', text)
-            self.assertNotIn('"old": true', text.lower())
-        finally:
-            path.unlink(missing_ok=True)
-
-    def test_invalid_json_shows_error(self):
-        url = reverse("projekt_file_edit_json", args=[self.file.pk])
-        resp = self.client.post(
-            url,
-            {"analysis_json": "{", "manual_analysis_json": "{}"},
-        )
-        self.assertEqual(resp.status_code, 200)
-        self.file.refresh_from_db()
-        self.assertEqual(
-            self.file.analysis_json, {"old": {"value": True, "editable": True}}
-        )
-
-    def test_question_review_saved(self):
-        url = reverse("projekt_file_edit_json", args=[self.anlage1.pk])
-        resp = self.client.post(
-            url,
-            {"q1_ok": "on", "q1_note": "Hinweis"},
-        )
-        self.assertRedirects(resp, reverse("projekt_detail", args=[self.projekt.pk]))
-        self.anlage1.refresh_from_db()
-        self.assertTrue(self.anlage1.question_review["1"]["ok"])
-        self.assertEqual(self.anlage1.question_review["1"]["note"], "Hinweis")
-
-    def test_question_review_extended_fields_saved(self):
-        url = reverse("projekt_file_edit_json", args=[self.anlage1.pk])
-        resp = self.client.post(
-            url,
-            {
-                "q1_status": "unvollst\u00e4ndig",
-                "q1_hinweis": "Fehlt",
-                "q1_vorschlag": "Mehr Infos",
-            },
-        )
-        self.assertRedirects(resp, reverse("projekt_detail", args=[self.projekt.pk]))
-        self.anlage1.refresh_from_db()
-        data = self.anlage1.question_review["1"]
-        self.assertEqual(data["status"], "unvollst\u00e4ndig")
-        self.assertEqual(data["hinweis"], "Fehlt")
-        self.assertEqual(data["vorschlag"], "Mehr Infos")
-
-    def test_question_review_prefill_from_analysis(self):
-        """Initialwerte stammen aus der automatischen Analyse."""
-        self.anlage1.question_review = None
-        self.anlage1.analysis_json = {
-            "questions": {
-                "1": {
-                    "answer": "A",
-                    "status": "ok",
-                    "hinweis": "H",
-                    "vorschlag": "V",
-                }
-            }
-        }
-        self.anlage1.save()
-
-        url = reverse("projekt_file_edit_json", args=[self.anlage1.pk])
-        resp = self.client.get(url)
-        form = resp.context["form"]
-        self.assertEqual(form.initial["q1_status"], "ok")
-        self.assertEqual(form.initial["q1_hinweis"], "H")
-        self.assertEqual(form.initial["q1_vorschlag"], "V")
-
-    def test_edit_page_has_mde(self):
-        url = reverse("projekt_file_edit_json", args=[self.file.pk])
-        resp = self.client.get(url)
-        self.assertContains(resp, "easymde.min.css")
 
 
 class Anlage2ReviewTests(NoesisTestCase):
@@ -2455,77 +1494,6 @@ class WorkerAnlage3VisionTests(NoesisTestCase):
         ) as mock_q:
             worker_run_anlage3_vision(self.projekt.pk, model_name="vision")
         self.assertEqual(mock_q.call_args[0][2], "vision")
-
-
-class GutachtenEditDeleteTests(NoesisTestCase):
-    def setUp(self):
-        self.user = User.objects.create_user("euser", password="pass")
-        self.client.login(username="euser", password="pass")
-        self.projekt = BVProject.objects.create(software_typen="A", beschreibung="x")
-        doc = Document()
-        doc.add_paragraph("Alt")
-        tmp = NamedTemporaryFile(delete=False, suffix=".docx")
-        doc.save(tmp.name)
-        tmp.close()
-        with open(tmp.name, "rb") as fh:
-            self.projekt.gutachten_file.save(
-                "g.docx", SimpleUploadedFile("g.docx", fh.read())
-            )
-        Path(tmp.name).unlink(missing_ok=True)
-        self.knowledge = SoftwareKnowledge.objects.create(
-            projekt=self.projekt,
-            software_name="A",
-            is_known_by_llm=True,
-        )
-        self.gutachten = Gutachten.objects.create(software_knowledge=self.knowledge, text="Alt")
-
-    def test_view_shows_content(self):
-        url = reverse("gutachten_view", args=[self.gutachten.pk])
-        resp = self.client.get(url)
-        self.assertContains(resp, "Alt")
-
-    def test_edit_updates_text(self):
-        url = reverse("gutachten_edit", args=[self.gutachten.pk])
-        resp = self.client.post(url, {"text": "Neu"})
-        self.assertRedirects(resp, reverse("gutachten_view", args=[self.gutachten.pk]))
-        self.gutachten.refresh_from_db()
-        self.assertEqual(self.gutachten.text, "Neu")
-
-    def test_edit_page_has_mde(self):
-        url = reverse("gutachten_edit", args=[self.gutachten.pk])
-        resp = self.client.get(url)
-        self.assertContains(resp, "easymde.min.css")
-
-    def test_delete_removes_file(self):
-        url = reverse("gutachten_delete", args=[self.gutachten.pk])
-        resp = self.client.post(url)
-        self.assertRedirects(resp, reverse("projekt_detail", args=[self.projekt.pk]))
-        self.assertFalse(Gutachten.objects.filter(pk=self.gutachten.pk).exists())
-
-
-class KnowledgeDescriptionEditTests(NoesisTestCase):
-    def setUp(self):
-        self.user = User.objects.create_user("kwuser", password="pass")
-        self.client.login(username="kwuser", password="pass")
-        self.projekt = BVProject.objects.create(software_typen="A", beschreibung="x")
-        self.knowledge = SoftwareKnowledge.objects.create(
-            projekt=self.projekt,
-            software_name="A",
-            is_known_by_llm=True,
-            description="Alt",
-        )
-
-    def test_edit_page_has_mde(self):
-        url = reverse("edit_knowledge_description", args=[self.knowledge.pk])
-        resp = self.client.get(url)
-        self.assertContains(resp, "easymde.min.css")
-
-    def test_edit_updates_description(self):
-        url = reverse("edit_knowledge_description", args=[self.knowledge.pk])
-        resp = self.client.post(url, {"description": "Neu"})
-        self.assertRedirects(resp, reverse("projekt_detail", args=[self.projekt.pk]))
-        self.knowledge.refresh_from_db()
-        self.assertEqual(self.knowledge.description, "Neu")
 
 
 class ProjektFileDeleteResultTests(NoesisTestCase):
@@ -2709,36 +1677,6 @@ class Anlage2ConfigSingletonTests(NoesisTestCase):
         self.assertEqual(Anlage2Config.objects.count(), 1)
 
 
-class AdminModelsViewTests(NoesisTestCase):
-    def setUp(self):
-        admin_group = Group.objects.create(name="admin")
-        self.user = User.objects.create_user("amodel", password="pass")
-        self.user.groups.add(admin_group)
-        self.client.login(username="amodel", password="pass")
-        self.cfg = LLMConfig.objects.create(
-            default_model="a",
-            gutachten_model="a",
-            anlagen_model="a",
-            available_models=["a", "b"],
-        )
-
-    def test_update_models(self):
-        url = reverse("admin_models")
-        resp = self.client.post(
-            url,
-            {
-                "default_model": "b",
-                "gutachten_model": "b",
-                "anlagen_model": "b",
-            },
-        )
-        self.assertRedirects(resp, url)
-        self.cfg.refresh_from_db()
-        self.assertEqual(self.cfg.default_model, "b")
-        self.assertEqual(self.cfg.gutachten_model, "b")
-        self.assertEqual(self.cfg.anlagen_model, "b")
-
-
 class Anlage1EmailTests(NoesisTestCase):
     def setUp(self):
         self.user = User.objects.create_user("emailer", password="pass")
@@ -2808,43 +1746,6 @@ class TileVisibilityTests(NoesisTestCase):
         resp = self.client.get(reverse("personal"))
         self.assertContains(resp, "<img", html=False)
 
-    def test_work_with_projekt_access(self):
-        UserTileAccess.objects.create(user=self.user, tile=self.projekt)
-        resp = self.client.get(reverse("work"))
-        self.assertContains(resp, "Projektverwaltung")
-        self.assertNotContains(resp, "TalkDiary")
-
-    def test_flag_reset_on_get(self):
-        self.cfg.models_changed = True
-        self.cfg.save(update_fields=["models_changed"])
-        url = reverse("admin_models")
-        self.client.get(url)
-        self.cfg.refresh_from_db()
-        self.assertFalse(self.cfg.models_changed)
-
-
-class TileAccessTests(NoesisTestCase):
-    def setUp(self):
-        work = Area.objects.get_or_create(slug="work", defaults={"name": "Work"})[0]
-        personal = Area.objects.get_or_create(
-            slug="personal", defaults={"name": "Personal"}
-        )[0]
-        self.talkdiary = Tile.objects.get_or_create(
-            slug="talkdiary",
-            defaults={
-                "name": "TalkDiary",
-                "url_name": "talkdiary_personal",
-            },
-        )[0]
-        self.talkdiary.areas.add(personal)
-        self.projekt = Tile.objects.get_or_create(
-            slug="projektverwaltung",
-            defaults={
-                "name": "Projektverwaltung",
-                "url_name": "projekt_list",
-            },
-        )[0]
-        self.projekt.areas.add(work)
 
     def _login(self, name: str) -> User:
         """Erzeugt einen Benutzer und loggt ihn ein."""
@@ -2998,74 +1899,6 @@ class RecordingDeleteTests(NoesisTestCase):
         self.assertTrue(Recording.objects.filter(pk=rec.pk).exists())
 
 
-class AdminAnlage1ViewTests(NoesisTestCase):
-    def setUp(self):
-        admin_group = Group.objects.create(name="admin")
-        self.user = User.objects.create_user("a1admin", password="pass")
-        self.user.groups.add(admin_group)
-        self.client.login(username="a1admin", password="pass")
-
-    def test_delete_question(self):
-        url = reverse("admin_anlage1")
-        questions = list(Anlage1Question.objects.all())
-        q = questions[0]
-        data = {}
-        for question in questions:
-            if question.id == q.id:
-                data[f"delete{question.id}"] = "on"
-            if question.parser_enabled:
-                data[f"parser_enabled{question.id}"] = "on"
-            if question.llm_enabled:
-                data[f"llm_enabled{question.id}"] = "on"
-            data[f"text{question.id}"] = question.text
-        resp = self.client.post(url, data)
-        self.assertRedirects(resp, url)
-        self.assertFalse(Anlage1Question.objects.filter(id=q.id).exists())
-        self.assertEqual(Anlage1Question.objects.count(), len(questions) - 1)
-
-    def _build_post_data(self, *, new=False, parser=True, llm=True):
-        """Hilfsfunktion zum Erstellen der POST-Daten."""
-        data = {}
-        for q in Anlage1Question.objects.all():
-            if q.parser_enabled:
-                data[f"parser_enabled{q.id}"] = "on"
-            if q.llm_enabled:
-                data[f"llm_enabled{q.id}"] = "on"
-            data[f"text{q.id}"] = q.text
-        if new:
-            data["new_text"] = "Neue Frage?"
-            if parser:
-                data["new_parser_enabled"] = "on"
-            if llm:
-                data["new_llm_enabled"] = "on"
-        return data
-
-    def test_add_new_question_with_flags(self):
-        url = reverse("admin_anlage1")
-        count = Anlage1Question.objects.count()
-        resp = self.client.post(
-            url, self._build_post_data(new=True, parser=True, llm=False)
-        )
-        self.assertRedirects(resp, url)
-        self.assertEqual(Anlage1Question.objects.count(), count + 1)
-        q = Anlage1Question.objects.order_by("-num").first()
-        self.assertEqual(q.text, "Neue Frage?")
-        self.assertTrue(q.parser_enabled)
-        self.assertFalse(q.llm_enabled)
-
-    def test_add_new_question_unchecked(self):
-        url = reverse("admin_anlage1")
-        count = Anlage1Question.objects.count()
-        resp = self.client.post(
-            url, self._build_post_data(new=True, parser=False, llm=False)
-        )
-        self.assertRedirects(resp, url)
-        self.assertEqual(Anlage1Question.objects.count(), count + 1)
-        q = Anlage1Question.objects.order_by("-num").first()
-        self.assertFalse(q.parser_enabled)
-        self.assertFalse(q.llm_enabled)
-
-
 class ModelSelectionTests(NoesisTestCase):
     def setUp(self):
         self.user = User.objects.create_user("modeluser", password="pass")
@@ -3121,90 +1954,6 @@ class ModelSelectionTests(NoesisTestCase):
             resp = self.client.post(url, {"model": "mf"})
         self.assertEqual(resp.status_code, 200)
         mock_func.assert_called_with(self.projekt.pk, model_name="mf")
-
-
-class CommandModelTests(NoesisTestCase):
-    def test_command_passes_model(self):
-        projekt = BVProject.objects.create(software_typen="A", beschreibung="x")
-        BVProjectFile.objects.create(
-            projekt=projekt,
-            anlage_nr=2,
-            upload=SimpleUploadedFile("a.txt", b"d"),
-            text_content="Text",
-        )
-        with patch("core.management.commands.check_anlage2.check_anlage2") as mock_func:
-            mock_func.return_value = {"ok": True}
-            call_command("check_anlage2", str(projekt.pk), "--model", "m3")
-        mock_func.assert_called_with(projekt.pk, model_name="m3")
-
-    def test_analyse_command_passes_model(self):
-        projekt = BVProject.objects.create(software_typen="A", beschreibung="x")
-        BVProjectFile.objects.create(
-            projekt=projekt,
-            anlage_nr=2,
-            upload=SimpleUploadedFile("b.txt", b"d"),
-            text_content="- Login",
-        )
-        with patch(
-            "core.management.commands.analyse_anlage2.analyse_anlage2"
-        ) as mock_func:
-            mock_func.return_value = {"missing": [], "additional": []}
-            call_command("analyse_anlage2", str(projekt.pk), "--model", "m4")
-        mock_func.assert_called_with(projekt.pk, model_name="m4")
-
-
-
-class Anlage2FunctionTests(NoesisTestCase):
-    def test_check_anlage2_functions_creates_result(self):
-        projekt = BVProject.objects.create(software_typen="A", beschreibung="x")
-        func = Anlage2Function.objects.create(name="Login")
-        llm_reply = json.dumps({
-            "technisch_verfuegbar": True,
-            "ki_beteiligung": False,
-        })
-        with patch("core.llm_tasks.query_llm", return_value=llm_reply):
-            data = check_anlage2_functions(projekt.pk)
-        res = Anlage2FunctionResult.objects.get(projekt=projekt, funktion=func)
-        self.assertTrue(res.technisch_verfuegbar)
-        self.assertIs(res.ki_beteiligung, False)
-        self.assertTrue(data[0]["technisch_verfuegbar"])
-        self.assertEqual(list(data[0].keys()), ["technisch_verfuegbar", "ki_beteiligung", "source", "funktion"])
-
-
-class CommandFunctionsTests(NoesisTestCase):
-    def test_functions_command_passes_model(self):
-        projekt = BVProject.objects.create(software_typen="A", beschreibung="x")
-        Anlage2Function.objects.create(name="Login")
-        with patch(
-            "core.management.commands.check_anlage2_functions.check_anlage2_functions"
-        ) as mock_func:
-            mock_func.return_value = []
-            call_command("check_anlage2_functions", str(projekt.pk), "--model", "m5")
-        mock_func.assert_called_with(projekt.pk, model_name="m5")
-
-
-
-
-
-
-
-class ProjektDetailAdminButtonTests(NoesisTestCase):
-    def setUp(self):
-        admin_group = Group.objects.create(name="admin")
-        self.admin = User.objects.create_user("padmin", password="pass")
-        self.admin.groups.add(admin_group)
-        self.user = User.objects.create_user("puser", password="pass")
-        self.projekt = BVProject.objects.create(software_typen="A", beschreibung="x")
-
-    def test_admin_user_sees_link(self):
-        self.client.login(username="padmin", password="pass")
-        resp = self.client.get(reverse("projekt_detail", args=[self.projekt.pk]))
-        self.assertContains(resp, reverse("admin_projects"))
-
-    def test_regular_user_hides_link(self):
-        self.client.login(username="puser", password="pass")
-        resp = self.client.get(reverse("projekt_detail", args=[self.projekt.pk]))
-        self.assertNotContains(resp, reverse("admin_projects"))
 
 
 class FunctionImportExportTests(NoesisTestCase):
@@ -3678,176 +2427,5 @@ class Anlage2ConfigViewTests(NoesisTestCase):
         self.assertRedirects(resp, url + "?tab=text")
         self.cfg.refresh_from_db()
         self.assertEqual(self.cfg.text_technisch_verfuegbar_true, ["ja", "okay"])
-
-
-class Anlage4ParserTests(NoesisTestCase):
-    def test_parse_table_and_regex(self):
-        cfg = Anlage4Config.objects.create(
-            table_columns=["zweck"],
-            regex_patterns=[r"Zweck: (.+)"]
-        )
-        doc = Document()
-        table = doc.add_table(rows=1, cols=2)
-        table.cell(0, 0).text = "Zweck"
-        table.cell(0, 1).text = "A"
-        tmp = NamedTemporaryFile(delete=False, suffix=".docx")
-        doc.save(tmp.name)
-        tmp.close()
-        with open(tmp.name, "rb") as fh:
-            upload = SimpleUploadedFile("t.docx", fh.read())
-        Path(tmp.name).unlink(missing_ok=True)
-        pf = BVProjectFile.objects.create(
-            projekt=BVProject.objects.create(software_typen="A"),
-            anlage_nr=4,
-            upload=upload,
-            text_content="Zweck: B",
-            anlage4_config=cfg,
-        )
-        items = parse_anlage4(pf)
-        self.assertEqual(items, ["A"])
-
-    def test_logs_table_detection(self):
-        cfg = Anlage4Config.objects.create(table_columns=["zweck"])
-        doc = Document()
-        table = doc.add_table(rows=1, cols=2)
-        table.cell(0, 0).text = "Zweck"
-        table.cell(0, 1).text = "A"
-        tmp = NamedTemporaryFile(delete=False, suffix=".docx")
-        doc.save(tmp.name)
-        tmp.close()
-        with open(tmp.name, "rb") as fh:
-            upload = SimpleUploadedFile("t.docx", fh.read())
-        Path(tmp.name).unlink(missing_ok=True)
-        pf = BVProjectFile.objects.create(
-            projekt=BVProject.objects.create(software_typen="A"),
-            anlage_nr=4,
-            upload=upload,
-            anlage4_config=cfg,
-        )
-        with self.assertLogs("anlage4_debug", level="DEBUG") as cm:
-            parse_anlage4(pf)
-        self.assertIn("table detected - 1 items", cm.output[0])
-
-    def test_negative_pattern(self):
-        cfg = Anlage4Config.objects.create(negative_patterns=["keine zwecke"])
-        pf = BVProjectFile.objects.create(
-            projekt=BVProject.objects.create(software_typen="A"),
-            anlage_nr=4,
-            upload=SimpleUploadedFile("x.docx", b""),
-            text_content="Keine Zwecke vorhanden",
-            anlage4_config=cfg,
-        )
-        self.assertEqual(parse_anlage4(pf), [])
-
-    def test_logs_free_text_detection(self):
-        cfg = Anlage4Config.objects.create(regex_patterns=[r"Zweck: (.+)"])
-        pf = BVProjectFile.objects.create(
-            projekt=BVProject.objects.create(software_typen="A"),
-            anlage_nr=4,
-            upload=SimpleUploadedFile("x.txt", b""),
-            text_content="Zweck: A",
-            anlage4_config=cfg,
-        )
-        with self.assertLogs("anlage4_debug", level="DEBUG") as cm:
-            parse_anlage4(pf)
-        self.assertIn("free text found - 1 items", cm.output[0])
-
-
-class AnalyseAnlage4Tests(NoesisTestCase):
-    def test_task_stores_json(self):
-        cfg = Anlage4Config.objects.create(prompt_template="Antwort:")
-        projekt = BVProject.objects.create(software_typen="A")
-        pf = BVProjectFile.objects.create(
-            projekt=projekt,
-            anlage_nr=4,
-            upload=SimpleUploadedFile("a.docx", b""),
-            text_content="Zweck: A",
-            anlage4_config=cfg,
-        )
-        with patch("core.llm_tasks.query_llm", return_value='{"text":"t","score":10,"begruendung":"b"}'):
-            data = analyse_anlage4(projekt.pk)
-        pf.refresh_from_db()
-        self.assertEqual(data["plausibility_score"]["value"], 10)
-        self.assertEqual(pf.analysis_json["plausibility_text"]["value"], "t")
-
-    def test_passes_config_to_parser(self):
-        cfg = Anlage4Config.objects.create(regex_patterns=[r"Zweck: (.+)"])
-        projekt = BVProject.objects.create(software_typen="A")
-        pf = BVProjectFile.objects.create(
-            projekt=projekt,
-            anlage_nr=4,
-            upload=SimpleUploadedFile("a.txt", b""),
-            text_content="Zweck: A",
-            anlage4_config=cfg,
-        )
-        with patch(
-            "core.llm_tasks.parse_anlage4", return_value=[]
-        ) as m_parse, patch("core.llm_tasks.query_llm", return_value="{}"):
-            analyse_anlage4(projekt.pk)
-        m_parse.assert_called_once_with(pf, cfg)
-
-
-class AnalyseAnlage4AsyncTests(NoesisTestCase):
-    def test_async_analysis_stores_results(self):
-        cfg = Anlage4Config.objects.create(regex_patterns=[r"Zweck: (.+)"])
-        projekt = BVProject.objects.create(software_typen="A")
-        pf = BVProjectFile.objects.create(
-            projekt=projekt,
-            anlage_nr=4,
-            upload=SimpleUploadedFile("a.txt", b""),
-            text_content="Zweck: A\nZweck: B",
-            anlage4_config=cfg,
-        )
-
-        def immediate(name, *args):
-            self.assertEqual(name, "core.llm_tasks.worker_anlage4_evaluate")
-            worker_anlage4_evaluate(*args)
-
-        with patch("core.llm_tasks.async_task", side_effect=immediate), patch(
-            "core.llm_tasks.query_llm",
-            return_value='{"text":"ok","score":5,"begruendung":"passt"}',
-        ):
-            analyse_anlage4_async(projekt.pk)
-
-        pf.refresh_from_db()
-        results = pf.analysis_json["zwecke"]
-        for item in results:
-            self.assertEqual(item["llm_result"]["text"], "ok")
-            self.assertEqual(item["llm_result"]["score"], 5)
-            self.assertEqual(item["llm_result"]["begruendung"], "passt")
-
-
-class Anlage4ReviewViewTests(NoesisTestCase):
-    def setUp(self):
-        self.user = User.objects.create_user("rev4", password="pass")
-        self.client.login(username="rev4", password="pass")
-        self.projekt = BVProject.objects.create(software_typen="A")
-        self.file = BVProjectFile.objects.create(
-            projekt=self.projekt,
-            anlage_nr=4,
-            upload=SimpleUploadedFile("a.txt", b""),
-            analysis_json={"zwecke": ["A", "B"]},
-        )
-
-    def test_post_saves_manual_review(self):
-        url = reverse("anlage4_review", args=[self.file.pk])
-        resp = self.client.post(
-            url,
-            {"item0_ok": "on", "item0_note": "gut", "item1_note": "schlecht"},
-        )
-        self.assertRedirects(resp, reverse("projekt_detail", args=[self.projekt.pk]))
-        self.file.refresh_from_db()
-        self.assertEqual(
-            self.file.manual_analysis_json,
-            {"0": {"ok": True, "note": "gut"}, "1": {"ok": False, "note": "schlecht"}},
-        )
-
-
-
-
-
-
-
-
